refactor: splitting interface and manager on another files
